Migrate OX Quiz Questions to DB (#691)
* ox quiz questions migrate

* formatting

* Update DatabaseOXQuizQuestion.cs
